ballast stuff, mostly for traj

+++ hootl/lasagna/PastaSim.h
@@ -30,7 +30,7 @@ public:
 	float sunset_dldt = 0;
 	SunsetPredictor sunpred;
 	unsigned int unix_offset;
-	unsigned int time = 0;				// sim time in ms
+	unsigned long time = 0;				// sim time in ms
 	float h;							// sim altitude in m
 	float v;							// sim velocity in m/s
 	float l;							// sim balloon lift in g
